refactor: 파일 유효성 검사 validateFile 유틸 함수로 분리 적용(#356)
